ml-service/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
pymysql==1.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
